rossmann-telegram-api/requirements.txt
Flask==2.0.1
json5==0.9.6
pandas==1.3.2
requests==2.26.0
seaborn==0.11.1
python-telegram-bot==13.7

